MFC: r299753 Fix fuse to use DIRECT_IO when required.
When a file is opened write-only and a partial block was written,
buffered I/O would try and read the whole block in. This would
result in a hung thread, since there was no open (fuse filehandle)
that allowed reading. This patch avoids the problem by forcing
DIRECT_IO for this case.
It also sets DIRECT_IO when the file system specifies the FN_DIRECTIO
flag in its reply to the open.


git-svn-id: https://svn.freebsd.org/base/stable/10@300916 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f