MFC r300258:
iic_rdwr_data->nmsgs is uint32_t, so limit the allowable number of messages to
prevent memory exhaustion and short allocations on 32-bit systems. Since
iicrdwr is intended to be a workalike of a Linux i2c-dev call, use the same
limit of 42 that Linux uses.

Also check the return value of copyin(9) to prevent unnecessary allocation in
the failure case.


git-svn-id: https://svn.freebsd.org/base/stable/10@300948 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f