Let <sched.h> define struct timespec.
POSIX 2004 doesn't require that this header defines struct timespec, but
does allow it. For FreeBSD, we would at least need a forward declaration
for sched_rr_get_interval().

POSIX 2008 is a bit more demanding, as it now requires that it is
declared. Remove the old forward declaration and include the _timespec.h
header. This should conform to both POSIX 2004 and 2008 now.


git-svn-id: https://svn.freebsd.org/base/head@301026 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f