MFC r299594
sfxge(4): avoid duplicate delivery of packets when changing multicast
mode with multicast chaining enabled

With multicast chaining, if e.g. a specific multicast filter is
inserted and the multicast mis-match filter is then inserted, both may
match a packet and cause it to be delivered.

Copy the behaviour of the Linux driver, which is to remove the old filters
first, on the basis that customers are more likely to be able to handle
drops than duplicates (see bug49178 comment 4).

Submitted by:   Mark Spender <mspender at solarflare.com>
Sponsored by:   Solarflare Communications, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/10@301342 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f