MFC r299595
sfxge(4): comment on when we assume multicast chaining is available

It's the same on Medford as Huntington.

Multicast chaining is not always on, even with Medford, as it's not
supported by low latency firmware.

Unlike the Linux driver, we don't need to support virtulization with
firmware released before support for multicast chaining was added.

Submitted by:   Mark Spender <mspender at solarflare.com>
Sponsored by:   Solarflare Communications, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/10@301343 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f