MFC r300008
sfxge(4): query and use current MTU if setting the MTU fails

This allows the driver to fall back to the largest usable MTU if a
user attempts to configure an unprivileged function with an MTU higher
than that of the attached port.

Submitted by:   Mark Spender <mspender at solarflare.com>
Sponsored by:   Solarflare Communications, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/10@301380 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f