MFC r301139
The (i < PROMPTLEN - 1) test added by r300442 in the code for the default
case of \c in the prompt format string is a no-op.  We already passed
this test at the top of the loop, and i has not yet been incremented in
this path.  Change this test to (i < PROMPTLEN - 2).

Reported by:	Coverity
CID:		1008328
Reviewed by:	cem
Differential Revision:	https://reviews.freebsd.org/D6552


git-svn-id: https://svn.freebsd.org/base/stable/10@301571 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f