Don't let ATF call basename() and dirname() in a non-standard way.
POSIX basename() and dirname() are allowed to overwrite the input
buffer. The advantage of implementing it that way is that the functions
can be thread-safe, as they don't store the result in a global buffer.

It looks like ATF wants to call basename() and dirname() in the
non-standard way where the argument is of type "const char *". This
change disables the calls to basename() and dirname(), which only seems
to affect the formatting of some rare debug/error messages.

Reviewed by:	ngie
Differential Revision:	https://reviews.freebsd.org/D6627


git-svn-id: https://svn.freebsd.org/base/head@301696 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f