MFC r299460: r299460 (by cem):
fsck_ffs: Don't overrun mount device buffer

Maybe this case is impossible.  Either way, when attempting to "/dev/"-prefix a
non-global device name, check that we do not overrun the f_mntfromname buffer.

In this case, truncating (with strlcpy or similar) would not be useful, since
the f_mntfromname result of getmntpt() is passed directly to open(2) later.

CID:		1006789


git-svn-id: https://svn.freebsd.org/base/stable/10@301782 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f